move compactmanager, store into single import
