+++ aiuser/context/compaction/__init__.py
@@ -1,0 +1,4 @@
+from aiuser.context.compaction.manager import CompactionManager
+from aiuser.context.compaction.store import CompactionStore
+
+__all__ = ["CompactionManager", "CompactionStore"]
