+++ aiuser/context/compaction/schema.py
